+++ src/main/java/com/sparta/oishitable/domain/common/kakao/service/KakaoService.java
@@ -1,10 +1,10 @@
-package com.sparta.oishitable.domain.auth.kakao.service;
+package com.sparta.oishitable.domain.common.kakao.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta.oishitable.domain.auth.kakao.dto.response.KakaoAccount;
-import com.sparta.oishitable.domain.auth.kakao.dto.response.KakaoInfo;
-import com.sparta.oishitable.domain.auth.kakao.dto.response.OAuthToken;
+import com.sparta.oishitable.domain.common.kakao.dto.response.KakaoAccount;
+import com.sparta.oishitable.domain.common.kakao.dto.response.KakaoInfo;
+import com.sparta.oishitable.domain.common.kakao.dto.response.OAuthToken;
 import com.sparta.oishitable.domain.common.auth.dto.response.AuthLoginResponse;
 import com.sparta.oishitable.domain.common.user.entity.User;
 import com.sparta.oishitable.domain.common.user.entity.UserRole;
